requirements.txt
tensorflow
tensorflow_datasets
numpy
tritonclient[all]
Pillow
